Refactor FlowActions so the tests pass. (Dependency on importing store, and through it, history, was causing issues.)
